refactor: 주석 제거
